requirements.txt
bs4
pandas
numpy
requests
python-dotenv
selenium
ipykernel
openai
langchain